+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.0] - 22.9.2025
+
+### Added
+- added interface to search,get requests
+- added interface to get the diff of a single request
+
+### Fixed
+- better language in in the spec files
+
 ## [0.1.1] - 19.9.2025
 
 ### Changed